Attempting to deserialize form alien JSON file.
Not working. Make sure that the class members are matching the json file
members, and that the members are public.

This is an attempt to load and deserialize a JSON file that was not
serialized by Json.NET. My theory is that as long as the member names
and types are the same, we will have no problems.