Added other types to the whitelist from JSON loading
The genreated worlds are fairly monsterous now that humanoids aren't
the only monster type loaded.

Noticed that many battles would go on infinitely because neither side
could score a casualty on the other.

Part of the issue was the HD calculation was out of whack.

Also added a "stalemate" condition to battles. When a battle goes beyond
10 rounds, each side makes an opposed morale check that determines the
outcome.